Changed all instances of fabsf to fabs so that compilation on venice works.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@125 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/ATrous/atrous_1d_reconstruct.cc
@@ -139,7 +139,7 @@ void atrous1DReconstruct(long &xdim, float *&input,float *&output, Param &par)
     if(par.isVerbose())     std::cout << "\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b";
 
   } while( (iteration==1) || 
-	   (fabsf(oldsigma-newsigma)/newsigma > reconTolerance) );
+	   (fabs(oldsigma-newsigma)/newsigma > reconTolerance) );
 
   if(par.isVerbose()) std::cout << "Completed "<<iteration<<" iterations. ";
